+++ thinglinks-modules/thinglinks-modules-link/src/main/java/com/mqttsnet/thinglinks/link/controller/product/ProductCommandsRequestsController.java
@@ -27,7 +27,7 @@ private ProductCommandsRequestsService productCommandsRequestsService;
 * @return 单条数据
 */
 @GetMapping("selectOne")
-public ProductCommandsRequests selectOne(Integer id) {
+public ProductCommandsRequests selectOne(Long id) {
 return productCommandsRequestsService.selectByPrimaryKey(id);
 }
